"Display error message on duplicate short URL"
Added functionality to display an error message on the index page if a user tries to shorten a URL that has already been shortened. Prior to this change, the application would return a string with the error but now it informs the user on the main page. This improves the user experience by providing feedback about the submission.