Add loading spinner to QR code image
Introduced a loading spinner which will appear when QR code image is being generated. This spinner will disappear once the image has finished loading. This was done to improve the user experience, providing visual feedback while the image is loading and give a clear indication when the image is ready for use.